added routes for backend; minor changes
