+++ server/database.js
@@ -1,0 +1,17 @@
+const mysql = require('mysql2');
+const db = mysql.createConnection({
+    user: 'root',
+    host: 'localhost',
+    password: null,
+    database: 'croma'
+});
+
+db.connect((err) => {
+    if (err) {
+      console.error('Error connecting to database:', err);
+      return;
+    }
+    console.log('Connected to database.');
+  });
+
+  exports.databaseConnection = db;
