+++ server/routes/admin.js
@@ -1,0 +1,6 @@
+const { Router } = require('express');
+
+const router = Router();
+const db = require('../database').databaseConnection;
+
+module.exports = router;
