+++ server/routes/announcement.js
@@ -1,0 +1,18 @@
+const { Router } = require('express');
+
+const router = Router();
+const db = require('../database').databaseConnection;
+
+router.post('/db/announcement', (req, res) => {
+    const q = 'INSERT INTO announcements (`announcement_title`, `announcement_body`) VALUES (?)'
+    const values = [
+      req.body.announcement_title,
+      req.body.announcement_body,]
+    
+  
+    db.query(q,[values], (err, data) => {
+      if(err) console.error('ERROR', err);
+    })
+  })
+
+module.exports = router;
