+++ server/routes/student.js
@@ -1,0 +1,7 @@
+const { Router } = require('express');
+
+const router = Router();
+const db = require('../database').databaseConnection;
+
+
+module.exports = router;
